feat: implementing cleanup service in progress

+++ src/main/java/com/tomcvt/cvtcaptcha/dtos/CIOParameters.java
@@ -1,5 +1,5 @@
 package com.tomcvt.cvtcaptcha.dtos;
 
-public record CIOParameters(float clickRadius) {
+public record CIOParameters(int clickRadius) {
     
 }
